move wait funcs to cluster
