+++ internal/testutil/acc/atlas.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/constant"
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/advancedcluster"
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/cluster"
 	"github.com/stretchr/testify/require"
 	"go.mongodb.org/atlas-sdk/v20250312007/admin"
 )
@@ -42,9 +42,7 @@ func createCluster(tb testing.TB, projectID, name string) string {
 	req := clusterReq(name, projectID)
 	_, _, err := ConnV2().ClustersApi.CreateCluster(tb.Context(), projectID, &req).Execute()
 	require.NoError(tb, err, "Cluster creation failed: %s, err: %s", name, err)
-	// TODO: TEMPORARY CHANGE, DON'T MERGE
-	// TODO: TEMPORARY CHANGE, DON'T MERGE
-	stateConf := advancedcluster.CreateStateChangeConfig(tb.Context(), ConnV2(), projectID, name, 1*time.Hour)
+	stateConf := cluster.CreateStateChangeConfig(tb.Context(), ConnV2(), projectID, name, 1*time.Hour)
 	_, err = stateConf.WaitForStateContext(tb.Context())
 	require.NoError(tb, err, "Cluster creation failed: %s, err: %s", name, err)
 
@@ -56,9 +54,7 @@ func deleteCluster(projectID, name string) {
 	if err != nil {
 		fmt.Printf("Cluster deletion failed: %s %s, error: %s", projectID, name, err)
 	}
-	// TODO: TEMPORARY CHANGE, DON'T MERGE
-	// TODO: TEMPORARY CHANGE, DON'T MERGE
-	stateConf := advancedcluster.DeleteStateChangeConfig(context.Background(), ConnV2(), projectID, name, 1*time.Hour)
+	stateConf := cluster.DeleteStateChangeConfig(context.Background(), ConnV2(), projectID, name, 1*time.Hour)
 	_, err = stateConf.WaitForStateContext(context.Background())
 	if err != nil {
 		fmt.Printf("Cluster deletion failed: %s %s, error: %s", projectID, name, err)
